refactor(components): 重构关于页面组件结构，将组件拆分到独立文件并统一导出
将原本集中在单个文件中的关于页面组件拆分为多个独立文件，并创建统一的导出文件
优化组件组织结构，提高代码可维护性和可读性
