+++ src/components/about/LoadingState.tsx
@@ -1,0 +1,12 @@
+// 加载状态组件
+export function LoadingState() {
+  return (
+    <div className="min-h-screen bg-gray-50 dark:bg-gray-900 py-4 sm:py-8 px-4">
+      <div className="max-w-6xl mx-auto">
+        <div className="text-center py-8">
+          <div className="text-lg text-gray-600 dark:text-gray-400">加载中...</div>
+        </div>
+      </div>
+    </div>
+  );
+}
