+++ src/components/about/index.ts
@@ -1,0 +1,10 @@
+// About页面组件统一导出
+export { Icon } from './Icon';
+export { HeroSection } from './HeroSection';
+export { FeaturesSection } from './FeaturesSection';
+export { InstructionsSection } from './InstructionsSection';
+export { VersionSection } from './VersionSection';
+export { ProjectsSection } from './ProjectsSection';
+export { ContactSection } from './ContactSection';
+export { FooterSection } from './FooterSection';
+export { LoadingState } from './LoadingState';
